Updating symptom menu component with symptomps of respective body parts and its corresponding file structure
